+++ frontend/src/passthrough/paginator/paginatorPassThrough.scss
