+++ frontend/src/passthrough/select/selectPassThrough.scss
@@ -1,0 +1,8 @@
+.p-select.fam-custom-select {
+  border-bottom: 2px solid #334155;
+
+  // remove only the sides you don’t want
+  border-top: 0;
+  border-left: 0;
+  border-right: 0;
+}
