+++ frontend/src/passthrough/select/selectPassThrough.ts
@@ -1,0 +1,6 @@
+
+export const SELECT_PASS_THROUGH = {
+  root: {
+    class: 'fam-custom-select'
+  }
+};
